+++ src/slicer.cpp
@@ -930,7 +930,7 @@ Slicer::Slicer(Mesh* mesh, const coord_t thickness, const size_t slice_layer_cou
 
 #pragma omp parallel for default(none) shared(mesh, layers_ref)
     // Use a signed type for the loop counter so MSVC compiles
-    for (int layer_nr = 0; layer_nr < layers_ref.size(); layer_nr++)
+    for (int layer_nr = 0; layer_nr < static_cast<int>(layers_ref.size()); layer_nr++)
     {
         layers_ref[layer_nr].makePolygons(mesh, layer_nr == 0);
     }
